Major Requested Fixes
Consolidated components: Merged the message protocol and
gmail_message_impl into their client counterparts.

    Simplified architecture: Reduced the number of core packages from
four to two.

    Refactored dependency injection: Introduced an explicit register()
function for implementations.

    Added project scaffolding: Created a new root README.md and an
.env.example file.

    Overhauled documentation: Significantly condensed and rewrote all
documentation for clarity.

    Updated tests: Aligned the entire test suite with the new,
simplified structure.
