+++ .env.example
@@ -1,0 +1,9 @@
+GMAIL_ACCESS_TOKEN=your-access-token
+GMAIL_REFRESH_TOKEN=your-refresh-token
+GMAIL_TOKEN_URI=https://oauth2.googleapis.com/token
+GMAIL_CLIENT_ID=your-client-id.apps.googleusercontent.com
+GMAIL_CLIENT_SECRET=your-client-secret
+GMAIL_SCOPES=https://www.googleapis.com/auth/gmail.modify
+GMAIL_UNIVERSE_DOMAIN=googleapis.com
+GMAIL_ACCOUNT=
+GMAIL_TOKEN_EXPIRY=YYYY-MM-DDTHH:MM:SSZ
