+++ src/gmail_client_impl/tests/test_gmail_message_edge_cases.py
@@ -1,24 +1,19 @@
-"""Tests for edge cases and error handling in gmail_message_impl.
-
-This module tests various edge cases, error conditions, and boundary
-scenarios to ensure robust error handling.
-"""
+"""Edge case tests for the colocated GmailMessage implementation."""
 
 import base64
 
-from gmail_message_impl._impl import GmailMessage
+from gmail_client_impl.message_impl import GmailMessage
 
 
 class TestEdgeCases:
     """Test cases for edge cases and error conditions."""
 
-    # Constants for magic values in tests
     VERY_LONG_SUBJECT_MIN_LENGTH = 1000
     VERY_LONG_BODY_MIN_LENGTH = 30000
 
     def test_extremely_large_message_id(self) -> None:
-        """Test with very long message ID."""
-        long_id = "x" * 1000  # Very long ID
+        """Test handling of extremely large message IDs."""
+        long_id = "x" * 1000
         simple_email = "Subject: Long ID Test\r\n\r\nBody"
         encoded_data = base64.urlsafe_b64encode(simple_email.encode()).decode()
 
@@ -27,7 +22,7 @@ class TestEdgeCases:
         assert msg.subject == "Long ID Test"
 
     def test_unicode_in_message_content(self) -> None:
-        """Test handling of Unicode characters in message content."""
+        """Test handling of Unicode characters in subject and body."""
         email_content = (
             "From: unicode@example.com\r\n"
             "Subject: 🎉 Unicode Test 测试 🌟\r\n"
@@ -38,13 +33,13 @@ class TestEdgeCases:
         encoded_data = base64.urlsafe_b64encode(email_content.encode("utf-8")).decode()
         msg = GmailMessage(msg_id="unicode123", raw_data=encoded_data)
 
-        assert "Unicode Test" in str(msg.subject)  # ASCII part should be preserved
+        assert "Unicode Test" in str(msg.subject)
         assert "こんにちは 世界! 🌍" in msg.body
         assert "Café naïve résumé" in msg.body
 
     def test_very_long_subject_line(self) -> None:
-        """Test with extremely long subject line."""
-        long_subject = "Very Long Subject " * 100  # ~1800 characters
+        """Test handling of very long subject lines."""
+        long_subject = "Very Long Subject " * 100
         email_content = f"Subject: {long_subject}\r\n\r\nShort body"
 
         encoded_data = base64.urlsafe_b64encode(email_content.encode()).decode()
@@ -54,8 +49,8 @@ class TestEdgeCases:
         assert len(msg.subject) > self.VERY_LONG_SUBJECT_MIN_LENGTH
 
     def test_very_long_message_body(self) -> None:
-        """Test with extremely long message body."""
-        long_body = "This is a very long message body. " * 1000  # ~35KB
+        """Test handling of very long message bodies."""
+        long_body = "This is a very long message body. " * 1000
         email_content = f"Subject: Long Body Test\r\n\r\n{long_body}"
 
         encoded_data = base64.urlsafe_b64encode(email_content.encode()).decode()
@@ -66,9 +61,9 @@ class TestEdgeCases:
         assert "This is a very long message body." in msg.body
 
     def test_malformed_headers(self) -> None:
-        """Test with malformed email headers."""
+        """Test handling of messages with malformed headers."""
         malformed_email = (
-            "From sender@example.com\r\n"  # Missing colon
+            "From sender@example.com\r\n"
             "Subject: Malformed Headers\r\n"
             "Invalid-Header-Without-Value\r\n"
             "\r\n"
@@ -78,23 +73,20 @@ class TestEdgeCases:
         encoded_data = base64.urlsafe_b64encode(malformed_email.encode()).decode()
         msg = GmailMessage(msg_id="malformed123", raw_data=encoded_data)
 
-        # Should handle gracefully
         assert msg.id == "malformed123"
         assert msg.body == "Invalid-Header-Without-Value\r\n\r\nBody content"
 
     def test_binary_data_in_raw_input(self) -> None:
-        """Test with binary data that's not valid email."""
-        binary_data = bytes(range(256))  # All possible byte values
+        """Test handling of raw data that includes binary content."""
+        binary_data = bytes(range(256))
         encoded_data = base64.urlsafe_b64encode(binary_data).decode()
 
-        # Should not crash, should handle gracefully
         msg = GmailMessage(msg_id="binary123", raw_data=encoded_data)
         assert msg.id == "binary123"
-        # Should use error fallback values
         assert msg.subject == "Error Parsing Message"
 
     def test_empty_raw_data(self) -> None:
-        """Test with empty raw data."""
+        """Test handling of completely empty raw data."""
         msg = GmailMessage(msg_id="empty123", raw_data="")
 
         assert msg.id == "empty123"
@@ -102,17 +94,16 @@ class TestEdgeCases:
         assert msg.from_ == "Unknown Sender"
 
     def test_whitespace_only_raw_data(self) -> None:
-        """Test with whitespace-only raw data."""
+        """Test raw data that is only whitespace characters."""
         whitespace_data = base64.urlsafe_b64encode(b"   \r\n\t  ").decode()
         msg = GmailMessage(msg_id="whitespace123", raw_data=whitespace_data)
 
         assert msg.id == "whitespace123"
-        # Should parse but may have empty fields
         assert isinstance(msg.subject, str)
         assert isinstance(msg.body, str)
 
     def test_non_ascii_message_id(self) -> None:
-        """Test with non-ASCII characters in message ID."""
+        """Test non-ASCII characters in message ID."""
         unicode_id = "msg_测试_🎉_123"
         simple_email = "Subject: Unicode ID Test\r\n\r\nBody"
         encoded_data = base64.urlsafe_b64encode(simple_email.encode()).decode()
@@ -122,8 +113,7 @@ class TestEdgeCases:
         assert msg.subject == "Unicode ID Test"
 
     def test_deeply_nested_multipart_message(self) -> None:
-        """Test with deeply nested multipart structure."""
-        # Create a complex nested structure
+        """Test deeply nested multipart messages."""
         nested_email = (
             "Content-Type: multipart/mixed; boundary=outer\r\n"
             "Subject: Nested Test\r\n"
@@ -155,7 +145,7 @@ class TestEdgeCases:
         assert "Plain text in nested structure" in msg.body
 
     def test_message_with_null_bytes(self) -> None:
-        """Test message containing null bytes."""
+        """Test message containing null bytes in headers and body."""
         email_with_nulls = "From: null@example.com\r\nSubject: Null Test\r\n\r\nBody with\x00null\x00bytes"
 
         encoded_data = base64.urlsafe_b64encode(email_with_nulls.encode("utf-8", errors="replace")).decode()
@@ -163,11 +153,10 @@ class TestEdgeCases:
 
         assert msg.id == "null123"
         assert msg.from_ == "null@example.com"
-        # Should handle null bytes gracefully
         assert "Body with" in msg.body
 
     def test_repeated_property_access(self) -> None:
-        """Test that property access is consistent on repeated calls."""
+        """Test that accessing properties multiple times yields consistent results."""
         email_content = (
             "From: repeat@example.com\r\nSubject: Repeat Test\r\nDate: Wed, 30 Jul 2025 10:30:00 +0000\r\n\r\nConsistent body"
         )
@@ -175,7 +164,6 @@ class TestEdgeCases:
         encoded_data = base64.urlsafe_b64encode(email_content.encode()).decode()
         msg = GmailMessage(msg_id="repeat123", raw_data=encoded_data)
 
-        # Call properties multiple times
         for _ in range(5):
             assert msg.id == "repeat123"
             assert msg.from_ == "repeat@example.com"
@@ -184,10 +172,9 @@ class TestEdgeCases:
             assert msg.body == "Consistent body"
 
     def test_message_with_only_headers_no_body(self) -> None:
-        """Test message with headers but no body."""
+        """Test message that contains only headers and no body."""
         headers_only = (
             "From: headeronly@example.com\r\nSubject: Headers Only\r\nDate: Wed, 30 Jul 2025 10:30:00 +0000\r\n\r\n"
-            # No body after the empty line
         )
 
         encoded_data = base64.urlsafe_b64encode(headers_only.encode()).decode()
@@ -197,4 +184,4 @@ class TestEdgeCases:
         assert msg.from_ == "headeronly@example.com"
         assert msg.subject == "Headers Only"
         assert msg.date == "07/30/2025"
-        assert msg.body == ""  # Should be empty string, not None
+        assert msg.body == ""
