+++ src/gmail_client_impl/tests/test_gmail_message_impl.py
@@ -1,13 +1,9 @@
-"""Tests for GmailMessage implementation.
-
-This module contains comprehensive tests for the GmailMessage class,
-covering message parsing, property access, error handling, and edge cases.
-"""
+"""Tests for GmailMessage implementation colocated with GmailClient."""
 
 import base64
 from email.message import EmailMessage
 
-from gmail_message_impl._impl import GmailMessage
+from gmail_client_impl.message_impl import GmailMessage
 
 
 class TestGmailMessage:
@@ -15,7 +11,6 @@ class TestGmailMessage:
 
     def test_basic_message_creation(self) -> None:
         """Test creating a GmailMessage with valid data."""
-        # Create a simple email message
         email_content = (
             "From: sender@example.com\r\n"
             "To: recipient@example.com\r\n"
@@ -25,13 +20,9 @@ class TestGmailMessage:
             "This is the message body."
         )
 
-        # Encode as Gmail would
         encoded_data = base64.urlsafe_b64encode(email_content.encode()).decode()
-
-        # Create GmailMessage instance
         message = GmailMessage(msg_id="test123", raw_data=encoded_data)
 
-        # Verify properties
         assert message.id == "test123"
         assert message.from_ == "sender@example.com"
         assert message.to == "recipient@example.com"
@@ -57,23 +48,18 @@ class TestGmailMessage:
 
     def test_multipart_message(self) -> None:
         """Test parsing a multipart message."""
-        # Create a multipart email
         email_msg = EmailMessage()
         email_msg["From"] = "sender@example.com"
         email_msg["To"] = "recipient@example.com"
         email_msg["Subject"] = "Multipart Test"
         email_msg["Date"] = "Wed, 30 Jul 2025 10:30:00 +0000"
 
-        # Add text part
         email_msg.set_content("This is the plain text body.")
-
-        # Add HTML part
         email_msg.add_alternative(
             "<html><body><p>This is the HTML body.</p></body></html>",
             subtype="html",
         )
 
-        # Encode the message
         raw_email = email_msg.as_bytes()
         encoded_data = base64.urlsafe_b64encode(raw_email).decode()
 
@@ -90,7 +76,6 @@ class TestGmailMessage:
         encoded_data = base64.urlsafe_b64encode(email_content.encode()).decode()
         message = GmailMessage(msg_id="baddate123", raw_data=encoded_data)
 
-        # Should fallback to raw date string
         assert message.date == "Invalid Date Format"
 
     def test_subject_without_encoding(self) -> None:
@@ -104,16 +89,12 @@ class TestGmailMessage:
 
     def test_message_with_attachment(self) -> None:
         """Test multipart message with attachment (should extract text body)."""
-        # Create multipart message with attachment
         email_msg = EmailMessage()
         email_msg["From"] = "sender@example.com"
         email_msg["Subject"] = "Message with Attachment"
         email_msg["Date"] = "Wed, 30 Jul 2025 10:30:00 +0000"
 
-        # Set main text content
         email_msg.set_content("This is the main message body.")
-
-        # Add an attachment
         email_msg.add_attachment(
             b"fake attachment data",
             maintype="application",
@@ -126,13 +107,11 @@ class TestGmailMessage:
 
         message = GmailMessage(msg_id="attach123", raw_data=encoded_data)
 
-        # Should extract the main text body, not the attachment
         assert "This is the main message body." in message.body
         assert "fake attachment data" not in message.body
 
     def test_message_body_encoding(self) -> None:
         """Test message body with non-UTF-8 encoding."""
-        # Create message with Latin-1 encoding
         body_text = "Café and naïve"
         email_content = (
             b"From: sender@example.com\r\nSubject: Encoding Test\r\nContent-Type: text/plain; charset=iso-8859-1\r\n\r\n"
@@ -141,10 +120,8 @@ class TestGmailMessage:
         encoded_data = base64.urlsafe_b64encode(email_content).decode()
         message = GmailMessage(msg_id="encoding123", raw_data=encoded_data)
 
-        # Should handle the encoding gracefully
         assert message.id == "encoding123"
         assert message.subject == "Encoding Test"
-        # Body should be decoded (may have replacement characters for encoding issues)
         assert len(message.body) > 0
 
     def test_message_without_plain_text_body(self) -> None:
@@ -153,7 +130,6 @@ class TestGmailMessage:
         email_msg["From"] = "sender@example.com"
         email_msg["Subject"] = "HTML Only"
 
-        # Add only HTML content, no plain text
         email_msg.set_content(
             "<html><body><h1>HTML Only Message</h1></body></html>",
             subtype="html",
@@ -164,5 +140,4 @@ class TestGmailMessage:
 
         message = GmailMessage(msg_id="htmlonly123", raw_data=encoded_data)
 
-        # Should indicate no plain text body found
         assert "<h1>HTML Only Message</h1>" in message.body
