--- src/gmail_message_impl/pyproject.toml
@@ -1,47 +1,0 @@
-[project]
-name = "gmail-message-impl"
-version = "0.1.0"
-description = "Add your description here"
-readme = "README.md"
-requires-python = ">=3.11"
-dependencies = [
-    "message",
-]
-
-[tool.ruff]
-line-length = 130  
-target-version = "py311"  # Adjust based on actual Python version
-extend = "../../pyproject.toml"
-
-[tool.ruff.lint]
-ignore = []
-
-[tool.uv.sources]
-message = { workspace = true }
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[project.optional-dependencies]
-test = [
-    "pytest>=7.0.0",
-    "pytest-mock>=3.10.0",
-]
-
-[tool.pytest.ini_options]
-pythonpath = [".", "src"]
-testpaths = ["tests", "src"]
-addopts = ["--cov", "--cov-report=term-missing"]
-
-[tool.coverage.run]
-source = ["src"]
-omit = ["*/tests/*", ]
-
-[tool.coverage.report]
-fail_under = 85 # Justification: A high threshold ensures most code is tested.
-exclude_lines = [
-    "pragma: no cover",
-    "raise NotImplementedError",
-    "if TYPE_CHECKING:",
-]
