--- src/gmail_message_impl/tests/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for gmail_message_impl package."""
