--- src/mail_client_api/tests/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for the mail client API package."""
