+++ src/mail_client_api/tests/test_mail_client_api.py
@@ -8,9 +8,7 @@ and serve as documentation for the expected API contracts.
 
 from unittest.mock import Mock
 
-from message import Message
-
-from mail_client_api import Client
+from mail_client_api import Client, Message
 
 
 def test_client_get_messages() -> None:
