--- src/message/pyproject.toml
@@ -1,19 +1,0 @@
-[project]
-name = "message"
-version = "0.1.0"
-description = "Add your description here"
-readme = "README.md"
-requires-python = ">=3.11"
-dependencies = []
-
-[tool.ruff]
-line-length = 100  # Default formatting width
-target-version = "py311"  # Adjust based on actual Python version
-extend = "../../pyproject.toml"
-
-[tool.ruff.lint]
-ignore = []
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
