--- src/message/tests/__init__.py
@@ -1,1 +1,0 @@
-"""Test for message API package."""
