--- tests/__init__.py
@@ -1,1 +1,0 @@
-"""E2E and Integration tests for Mail Client."""
