--- tests/e2e/__init__.py
@@ -1,6 +1,0 @@
-"""End-to-End tests package.
-
-This package contains E2E tests that verify the complete application
-workflow from a user's perspective. These tests interact with real
-external services and test the full system integration.
-"""
