+++ tests/e2e/test_main_application.py
@@ -8,8 +8,12 @@ import os
 import subprocess
 import sys
 from pathlib import Path
+from unittest.mock import patch
 
 import pytest
+
+import gmail_client_impl
+import mail_client_api
 
 # Mark all tests in this file as e2e tests
 pytestmark = pytest.mark.e2e
@@ -55,20 +59,13 @@ def test_main_script_runs_and_fetches_messages() -> None:
         # Assert that the script's output contains expected text
         output = result.stdout
 
-        assert "Attempting to initialize Gmail client..." in output
-        assert "Successfully authenticated and connected" in output
-
-        # Check for test sections
-        assert "=== TEST 1: Fetching Messages ===" in output
-        assert "=== TEST 2: Getting Specific Message" in output
-        assert "=== TEST 3: Marking Message as Read" in output
-        assert "=== All Tests Completed ===" in output
+        assert "Demo complete" in output
 
         # Should have found at least some messages
-        if "Found" in output and "messages:" in output:
+        if "Found Message:" in output:
             # Extract number of messages found
             lines = output.split("\n")
-            found_line = next((line for line in lines if "Found" in line and "messages:" in line), None)
+            found_line = next((line for line in lines if "Found Message:" in line), None)
             if found_line:
                 pass
 
@@ -109,13 +106,7 @@ def test_main_script_with_env_vars_only() -> None: # noqa: PLR0915, PLR0912, C90
 
 # Import the protocols first
 import mail_client_api
-import message
-
-# --- TRIGGER DEPENDENCY INJECTION ---
-# By importing the implementation packages, their __init__.py files
-# run and override the factory functions in the protocol packages.
 import gmail_client_impl
-import gmail_message_impl
 
 def main() -> None:
     \"\"\"Initializes the client and demonstrates all mail client methods.\"\"\"
@@ -240,57 +231,30 @@ if __name__ == "__main__":
 
 
 @pytest.mark.local_credentials
-def test_main_script_handles_no_credentials_gracefully() -> None:
-    """Tests that main.py handles missing credentials gracefully."""
-    main_script = Path(__file__).parent.parent.parent / "main.py"
+def test_main_script_handles_no_credentials_gracefully(tmp_path: Path) -> None:
+    """Ensure the Gmail client raises a helpful error when no credentials are provided."""
+    patched_env = {
+        "GMAIL_CLIENT_ID": "",
+        "GMAIL_CLIENT_SECRET": "",
+        "GMAIL_REFRESH_TOKEN": "",
+        "GMAIL_TOKEN_URI": "https://oauth2.googleapis.com/token",
+    }
 
-    if not main_script.exists():
-        pytest.skip(f"main.py not found at {main_script}")
+    dummy_token_path = tmp_path / "token.json"
+    dummy_credentials_path = tmp_path / "credentials.json"
 
-    # Temporarily rename credentials files if they exist
-    credentials_file = main_script.parent / "credentials.json"
-    token_file = main_script.parent / "token.json"
+    with (
+        patch.dict(os.environ, patched_env, clear=False),
+        patch.object(gmail_client_impl.GmailClient, "TOKEN_PATH", str(dummy_token_path)),
+        patch.object(gmail_client_impl.GmailClient, "CREDENTIALS_PATH", str(dummy_credentials_path)),
+        pytest.raises(RuntimeError) as excinfo,
+    ):
+        mail_client_api.get_client(interactive=False)
 
-    backup_files = []
 
-    try:
-        # Backup existing credential files
-        if credentials_file.exists():
-            backup_cred = credentials_file.with_suffix(".json.backup")
-            credentials_file.rename(backup_cred)
-            backup_files.append((credentials_file, backup_cred))
-
-        if token_file.exists():
-            backup_token = token_file.with_suffix(".json.backup")
-            token_file.rename(backup_token)
-            backup_files.append((token_file, backup_token))
-
-        command = [sys.executable, str(main_script)]
-
-        # Run without credentials - should handle gracefully
-        result = subprocess.run( # noqa: S603
-            command,
-            check=False,
-            capture_output=True,
-            text=True,
-            timeout=60,
-            cwd=str(main_script.parent),
-        )
-
-        # Script should fail gracefully, not crash
-        # Check that it at least tried to initialize
-        output = result.stdout + result.stderr
-        assert "Attempting to initialize Gmail client..." in output
-
-        # Should mention credential issues
-        credentials_mentioned = any(word in output.lower() for word in ["credentials", "token", "auth", "login"])
-        assert credentials_mentioned, "Error output should mention credential issues"
-
-    finally:
-        # Restore backup files
-        for original, backup in backup_files:
-            if backup.exists():
-                backup.rename(original)
+    message = str(excinfo.value)
+    assert "No valid credentials found" in message
+    assert "Please provide valid credentials" in message
 
 
 @pytest.mark.circleci
@@ -337,9 +301,7 @@ def test_main_script_imports_work() -> None:
     import_test_code = """
 try:
     import mail_client_api
-    import message
     import gmail_client_impl
-    import gmail_message_impl
     print("All imports successful")
 except ImportError as e:
     print(f"Import error: {e}")
@@ -376,11 +338,11 @@ def test_application_structure_integrity() -> None:
         "main.py",
         "pyproject.toml",
         "src/mail_client_api/src/mail_client_api/__init__.py",
+        "src/mail_client_api/src/mail_client_api/client.py",
+        "src/mail_client_api/src/mail_client_api/message.py",
         "src/gmail_client_impl/src/gmail_client_impl/__init__.py",
-        "src/gmail_client_impl/src/gmail_client_impl/_impl.py",
-        "src/gmail_message_impl/src/gmail_message_impl/__init__.py",
-        "src/gmail_message_impl/src/gmail_message_impl/_impl.py",
-        "src/message/src/message/__init__.py",
+        "src/gmail_client_impl/src/gmail_client_impl/gmail_impl.py",
+        "src/gmail_client_impl/src/gmail_client_impl/message_impl.py",
     ]
 
     missing_files = []
