--- tests/integration/__init__.py
@@ -1,6 +1,0 @@
-"""Integration tests package.
-
-This package contains integration tests that verify how multiple components
-work together. These tests may interact with real external services but
-should be faster than E2E tests.
-"""
